+++ src/de.secucard.connect/client-node-environment.js
@@ -44,8 +44,11 @@ export const ClientNodeEnvironment = {
         General.StoreService,
         General.TransactionService,
 
-        Loyalty.ActionActionService,
+        Loyalty.ActionService,
         Loyalty.ActionProfileService,
+        Loyalty.ActionCampaignService,
+        Loyalty.ActionMessageService,
+        Loyalty.ActionConfigService,
         Loyalty.BeaconService,
         Loyalty.CardGroupService,
         Loyalty.CardService,
@@ -55,13 +58,17 @@ export const ClientNodeEnvironment = {
         Loyalty.MerchantCardService,
         Loyalty.ProgramService,
         Loyalty.ProgramSpecialService,
+        Loyalty.ReportService,
         Loyalty.SaleService,
         Loyalty.StoreGroupService,
+        Loyalty.TransactionService,
 
         Payment.ContainerService,
         Payment.ContractService,
         Payment.CustomerService,
+        Payment.SecupayCreditcardService,
         Payment.SecupayDebitService,
+        Payment.SecupayInvoiceService,
         Payment.SecupayPrepayService,
         Payment.TransactionService,
         
@@ -118,8 +125,11 @@ export const ServiceMap = {
         Transactions: General.TransactionService.Uid
     },
     Loyalty: {
+        ActionCampaigns: Loyalty.ActionCampaignService.Uid,
+        ActionConfigs: Loyalty.ActionConfigService.Uid,
+        ActionMessages: Loyalty.ActionMessageService.Uid,
         ActionProfiles: Loyalty.ActionProfileService.Uid,
-        ActionActions: Loyalty.ActionActionService.Uid,
+        Actions: Loyalty.ActionService.Uid,
         Beacons: Loyalty.BeaconService.Uid,
         CardGroups: Loyalty.CardGroupService.Uid,
         Cards: Loyalty.CardService.Uid,
@@ -129,14 +139,18 @@ export const ServiceMap = {
         MerchantCards: Loyalty.MerchantCardService.Uid,
         Programs: Loyalty.ProgramService.Uid,
         ProrgamSpecials: Loyalty.ProgramSpecialService.Uid,
+        Reports: Loyalty.ReportService.Uid,
         Sales: Loyalty.SaleService.Uid,
-        StoreGroups: Loyalty.StoreGroupService.Uid
+        StoreGroups: Loyalty.StoreGroupService.Uid,
+        Transactions: Loyalty.TransactionService.Uid
     },
     Payment: {
         Containers: Payment.ContainerService.Uid,
         Contracts: Payment.ContractService.Uid,
         Customers: Payment.CustomerService.Uid,
+        SecupayCreditcards: Payment.SecupayCreditcardService.Uid,
         SecupayDebits: Payment.SecupayDebitService.Uid,
+        SecupayInvoices: Payment.SecupayInvoiceService.Uid,
         SecupayPrepays: Payment.SecupayPrepayService.Uid,
         Transactions: Payment.TransactionService.Uid
     },