+++ src/de.secucard.connect/product/payment/contract-service.js
@@ -1,37 +1,30 @@
-/*
- Copyright 2015 hp.weber GmbH & Co secucard KG (www.secucard.com)
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
- */
-import {ProductService} from '../product-service'
+import {ProductService} from '../product-service';
 
 export class ContractService extends ProductService {
 
-    constructor() {
-        super()
-    }
+  constructor() {
+    super();
+  }
 
-    getEndpoint() {
-        return ['payment', 'contracts'];
-    }
+  getEndpoint() {
+    return ['payment', 'contracts'];
+  }
 
-    getEventTargets() {
-        return [];
-    }
+  getEventTargets() {
+    return [];
+  }
 
-    clone(contractId, cloneParams) {
-        return this.execute(contractId, 'clone');
-    }
+  clone(contractId, cloneParams) {
+    return this.execute(contractId, 'clone', null, cloneParams);
+  }
 
-    cloneMine(cloneParams) {
-        return this.clone('me', cloneParams);
-    }
+  cloneMine(cloneParams) {
+    return this.clone('me', cloneParams);
+  }
+
+  createSubContract(createSubContractRequest) {
+    return this.execute('me', 'requestId', null, createSubContractRequest)
+  }
 
 }