+++ src/de.secucard.connect/product/payment/model/basket.js
@@ -1,0 +1,18 @@
+/*
+ * SDK-Generator version: 1.1.0
+ */
+export class Basket {
+  constructor(ean, tax, price, articleNumber, name, total, quantity, itemType, contractId, model, apikey) {
+    this['ean'] = ean;
+    this['tax'] = tax;
+    this['price'] = price;
+    this['article_number'] = articleNumber;
+    this['name'] = name;
+    this['total'] = total;
+    this['quantity'] = quantity;
+    this['item_type'] = itemType;
+    this['contract_id'] = contractId;
+    this['model'] = model;
+    this['apikey'] = apikey;
+  }
+}