+++ src/de.secucard.connect/product/payment/model/create-sub-contract-response.js
@@ -1,0 +1,11 @@
+/*
+ * SDK-Generator version: 1.1.0
+ */
+export class CreateSubContractResponse {
+  constructor(apikey, contract, payinAccount, contractId) {
+    this['apikey'] = apikey;
+    this['contract'] = contract;
+    this['payin_account'] = payinAccount;
+    this['contract_id'] = contractId;
+  }
+}