+++ src/de.secucard.connect/product/payment/model/customer.js
@@ -1,22 +1,13 @@
 /*
- Copyright 2015 hp.weber GmbH & Co secucard KG (www.secucard.com)
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
+ * SDK-Generator version: 1.1.0
  */
 export class Customer {
-
-    constructor(merchant, contact, created, updated, contract) {
-        this.merchant = merchant;
-        this.contact = contact;
-        this.created = created;
-        this.updated = updated;
-        this.contract = contract;
-    }
-
+  constructor(merchant, contact, created, updated, contract, merchantCustomerId) {
+    this['merchant'] = merchant;
+    this['contact'] = contact;
+    this['created'] = created;
+    this['updated'] = updated;
+    this['contract'] = contract;
+    this['merchant_customer_id'] = merchantCustomerId;
+  }
 }