+++ src/de.secucard.connect/product/payment/model/opt-data.js
@@ -1,0 +1,9 @@
+/*
+ * SDK-Generator version: 1.1.0
+ */
+export class OptData {
+  constructor(hasAcceptedDisclaimer, hideDisclaimer) {
+    this['has_accepted_disclaimer'] = hasAcceptedDisclaimer;
+    this['hide_disclaimer'] = hideDisclaimer;
+  }
+}