+++ src/de.secucard.connect/product/product-service.js
@@ -1,14 +1,3 @@
-/*
- Copyright 2015 hp.weber GmbH & Co secucard KG (www.secucard.com)
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
- */
 import {Channel} from '../net/channel';
 import EE from 'eventemitter3';
 
@@ -117,6 +106,7 @@ export class ProductService {
      * Would invoke for example: GET /targetType/objectId .<br/>
      * An exception is thrown if the id is unknown.
      * @param id
+     * @param queryParams
      * @param options
      * @return {Promise}
      */
@@ -133,6 +123,17 @@ export class ProductService {
         return this._request(Channel.METHOD.GET, params, options);
 
     }
+    
+    generateRetrieveUrl(id, queryParams, options) {
+        let params = {
+            endpoint: this.getEndpoint(),
+            objectId: id,
+            queryParams: queryParams,
+            options: options
+        };
+        
+        return this._generateUrl(Channel.METHOD.GET, params, options);
+    }
 
     /**
      * Retrieves a promise for a single object (resource) of a given type, never null<br/>
@@ -141,6 +142,7 @@ export class ProductService {
      * @param id
      * @param action
      * @param actionArg
+     * @param options
      * @return {Promise}
      */
     retrieveWithAction(id, action, actionArg, options) {
@@ -155,6 +157,19 @@ export class ProductService {
 
         return this._request(Channel.METHOD.GET, params, options);
 
+    }
+    
+    generateRetrieveWithActionUrl(id, action, actionArg, options) {
+        
+        let params = {
+            endpoint: this.getEndpoint(),
+            objectId: id,
+            action: action,
+            actionArg: actionArg,
+            options: options
+        };
+        
+        return this._generateUrl(Channel.METHOD.GET, params, options);
     }
 
     /**
@@ -177,6 +192,17 @@ export class ProductService {
 
         return this._request(Channel.METHOD.GET, params, options);
 
+    }
+    
+    generateRetrieveListUrl(queryParams, options) {
+        
+        let params = {
+            endpoint: this.getEndpoint(),
+            queryParams: queryParams,
+            options: options
+        };
+        
+        return this._generateUrl(Channel.METHOD.GET, params, options);
     }
 
     /**
@@ -273,7 +299,6 @@ export class ProductService {
      *
      * @param id   Id of the resource to delete.
      * @param options        will be used to determine actual target destination.
-     * @param callback   The callback for async invocation.
      */
 
     remove(id, options) {
@@ -385,6 +410,20 @@ export class ProductService {
 
         return this.getChannel(options.channelConfig).request(method, params);
 
+    }
+    
+    _generateUrl(method, params, options) {
+        
+        if (options == null) {
+            options = this.getServiceDefaultOptions();
+        }
+
+        if (params.options == null) {
+            params.options = options;
+        }
+        
+        return this.getChannel([Channel.REST]).generateUrl(method, params);
+        
     }
 
 }